requirements.txt
telethon
pyfuse3
typing_extensions
pytest
greenback
aiofiles
pytest-asyncio
pyyaml
tqdm